+++ src/js/moss.js
@@ -1,6 +1,6 @@
 /*!
  * moss-ui - The front-end UI framework with Vue.js and SCSS.
- * @version v0.7.1
+ * @version v0.7.2
  * @link https://github.com/archco/moss-ui
  * @license MIT
  */