+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+## [0.7.2][] - 2018-07-21
+
 ### Added
 
 - Add scss mixin `reset-link`. #140
@@ -337,7 +339,8 @@ Remove deprecated in v0.6.x #104
 - Remove unnecessary part "aside.scss". #60
 - Remove deprecated scss mixin, `flex-container()`. instead use `flex-container-align`. #52
 
-[Unreleased]: https://github.com/archco/moss-ui/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/archco/moss-ui/compare/v0.7.2...HEAD
+[0.7.2]: https://github.com/archco/moss-ui/compare/v0.7.0...v0.7.2
 [0.7.0]: https://github.com/archco/moss-ui/compare/v0.6.5...v0.7.0
 [0.6.5]: https://github.com/archco/moss-ui/compare/v0.6.4...v0.6.5
 [0.6.4]: https://github.com/archco/moss-ui/compare/v0.6.3...v0.6.4